back-end/requirements.txt
starlette
databases
sqlalchemy
uvicorn
psycopg2-binary
asyncpg
python-dotenv

